fix: harden auth, file access, and AI task safety

+++ backend/app/models/tenant.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import Boolean, DateTime, Integer, JSON, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -16,8 +17,8 @@ class Tenant(Base, TimestampMixin):
     type: Mapped[str] = mapped_column(String(30), nullable=False, default="organization")
     status: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
     subscription_expire_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
-    balance: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False, default=0)
-    ai_monthly_budget_limit: Mapped[float | None] = mapped_column(Numeric(12, 2))
+    balance: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False, default=0)
+    ai_monthly_budget_limit: Mapped[Decimal | None] = mapped_column(Numeric(12, 2))
     ai_budget_degrade_model: Mapped[str | None] = mapped_column(String(100))
     ai_prompt_settings: Mapped[dict | None] = mapped_column(JSON, default=dict)
     ai_provider_settings: Mapped[dict | None] = mapped_column(JSON, default=dict)
